requirements.txt
ansible==4.3.0
ansible-core==2.11.3
ansible-lint==5.1.2
arrow==1.1.1
attrs==21.2.0
backports.entry-points-selectable==1.1.0
bcrypt==3.2.0
binaryornot==0.4.4
bracex==2.1.1
Cerberus==1.3.2
certifi==2021.5.30
cffi==1.14.6
cfgv==3.3.0
chardet==4.0.0
charset-normalizer==2.0.4
click==8.0.1
click-help-colors==0.9.1
colorama==0.4.4
commonmark==0.9.1
cookiecutter==1.7.3
cryptography==3.4.7
distlib==0.3.2
distro==1.6.0
docker==5.0.0
enrich==1.2.6
filelock==3.0.12
flake8==3.9.2
identify==2.2.12
idna==3.2
iniconfig==1.1.1
Jinja2==3.0.1
jinja2-time==0.2.0
MarkupSafe==2.0.1
mccabe==0.6.1
molecule==3.4.0
molecule-docker==0.2.4
nodeenv==1.6.0
packaging==21.0
paramiko==2.7.2
pathspec==0.9.0
platformdirs==2.2.0
pluggy==0.13.1
poyo==0.5.0
pre-commit==2.13.0
py==1.10.0
pycodestyle==2.7.0
pycparser==2.20
pyflakes==2.3.1
Pygments==2.9.0
PyNaCl==1.4.0
pyparsing==2.4.7
pytest==6.2.4
pytest-testinfra==6.4.0
python-dateutil==2.8.2
python-slugify==5.0.2
PyYAML==5.4.1
requests==2.26.0
resolvelib==0.5.4
rich==10.7.0
ruamel.yaml==0.17.10
ruamel.yaml.clib==0.2.6
selinux==0.2.1
six==1.16.0
subprocess-tee==0.3.2
tenacity==8.0.1
text-unidecode==1.3
toml==0.10.2
urllib3==1.26.6
virtualenv==20.7.0
wcmatch==8.2
websocket-client==1.1.0
yamllint==1.26.2
